fix(app): Do not confirm camera preferences if camera is disconnected from robot (#20178)
Closes RQA-4839

It's possible for users to confirm their camera preferences and "enable" the camera even if it's disconnected. There are no dire consequences for doing so, but we have room to improve this behavior on the UX side. There are a few copy errors in which we don't properly surface camera state that needed updating along the same lines.