fix(robot-server): delete input/output data file associations when removing runs (#20185)
Closes RABR-839

Currently, if a run is started and the 21st run is a CSV protocol, the run cannot start.

Run deletion is failing with a FOREIGN KEY constraint error when auto-deleting old runs to make room for new ones. The error occurred because run_store.remove() is not deleting associations in the input_data_files_table and output_data_files_table before attempting to delete the run record in the RunAutoDeleter.

To fix, we just add deletion of input and output data file associations in RunStore.remove() before deleting the run record itself. This ensures all foreign key constraints are satisfied during the deletion cascade.

Note that there is currently a 1:1 relationship between input files:runs and output files: runs, and we may have to revisit this logic in the future.