docs: quick transfer return tips update (#20180)
# Overview

Adds a sentence about return tips with the release of robot software
v8.8. This is in the Toucscreen section, Quick Transfer doc.

Sandbox: 
-
https://sandbox.docs.opentrons.com/docs-quick-transfer-update/flex/touchscreen/quick-transfer/#pipettes-and-tips

RTC-888

## Test Plan and Hands on Testing

N/A not much to test.

## Changelog

Revises "Pipettes and tips" section, specifically the 2nd paragraph, in
`quick-transfer.md`. Adds a sentence to reflect that v8.8 makes return
tips possible in a quick transfer.

## Review requests

This is a brief sentence. Is it sufficient or are we missing something?

## Risk assessment

Low.