feat(hardware-testing): add cavity testing option to the grav protocol (#20189)
# Overview
When we create a new batch of tips we use this script to create a
profile of each tip cavities performance relative to each other. The new
option essentially just runs 15 samples of each of the 16 tip cavities
in the injection mold in one run of the gravimetric script to compare
the results.