fix(protocol-designer): better error messages when can't multi-dispense (#20190)
# Overview

Previously, we would display the error `Multi dispense values not found
for liquid class` if a Transfer step had the Distribute pipette path
selected but we cannot fit more than 1 well's volume into the tip. (The
form prevents you from selecting Distribute in that case, but older
protocols could be saved with Distribute selected, such as in
AUTH-2511.)

The error doesn't make much sense and is confusing to the user, implying
that something is wrong with the liquid class definition. This PR
changes the error message to say that they can't use Distribute because
the volume is too high.

Also, we were using the same error message for **Consolidate** when we
couldn't fit more than 1 well's volume into the tip. There, the message
really doesn't make sense, because the problem has nothing to do with
`Multi dispense` at all.

## Test Plan and Hands on Testing

Tried it locally. Here's what the error message looks like now for the
protocol in the ticket:
<img width="635" height="61" alt="image" src="https://github.com/user-attachments/assets/f0a985db-81ea-49eb-82f4-9faa595feb3f" />

## Risk assessment

Low.