fix(api): bail early on module gcode parse errors (#20194)
We were waiting for multiple acks from certain commands, but if those
commands fail at the module's gcode parser stage (i.e. because the
module isn't updated and doesn't handle that gcode) then we'll never get
a second ack and we'll just wait a really long time. This is ultimately
fine in that it doesn't fail the protocol, but it makes anything that
uses modules that have this problem take a really really long time.

## testing
- [x] on a robot with a module that doesn't handle M411, check that the
module polls don't take 60 seconds to complete

Closes RABR-843