fix(api): fix emulation proxy teardown
We have these proxy servers, and they just hang when tearing them down
sometimes. Well, they used to; in 3.12, they hang when tearing them down
every time.

The reason is that the asyncio server code doesn't handle being shutdown
in the presence of connected clients very well. Previously, it might
hang; in 3.12, it will always hang because there's some code that
catches cancellation and tries to wait for a clean shutdown that will
never happen.

In <3.13 there's also no way to properly shut down clients in the public
api.

One workaround, which is awful, and which we're doing, is to just cancel
the server twice, waiting a little in between so that we can get the
server into that state where it's waiting a bit.

Luckily this is not really used in production code.