feat(protocol-designer): Update Shift-click step selection to work with concurrent Thermocycler profiles (#20193)
## Overview

Protocol Designer lets the user Shift-click to select a range of steps
in the timeline. Now that the timeline has nesting (concurrent
Thermocycler profiles), we need to update that logic to account for some
hazards when the range crosses the boundary of a nested group. This PR
does that.

Closes EXEC-2060. See that ticket for a more full example of the problem
we're solving.

## Changelog

* Update the logic in `getShiftSelectedSteps()` to avoid selecting steps
that are not "user-visible"—i.e., the hidden "wait for profile to
complete" step that implicitly comes at the end of every Thermocycler
profile. Some of the other step CRUD actions are already doing the same
thing.
* Various small renames for clarity.
* Delete `SelectMultipleStepsForGroupAction`, which was unused (never
created, nor used in any reducer).