fix(api): Correctly update engine state during tracking aspirate/dispense (#20195)
# Overview
We were not appending the aspirate/dispense result to the move result.
This means that the engine's state was not getting updated with the
current labware, so the *_while_tracking commands were not updating the
pipettes current location.

This means that if you did an action in labware_1, then did a
aspirate/dispense_while_tracking in labware_2 and then moved back to
labware_1 the engine would think that the pipettes current location was
still in labware 1 and it would calculate the safe Z movement and
properly use waypoints to travel back to labware_1.