fix: enum fixes for 3.12
We heavily used `class MyStringEnum(str, Enum)` to create string
enumerators that, when put in format strings, just result in their
value. This is because the __format__ dunder of these enums was
formatting them differently than other kinds of enums. In 3.11, this
became consistent, and so we need to use the new class StrEnum - which
has the proper behavior - instead. If StrEnum doesn't exist, we'll use a
weird override.

Also we really need to use __new__ overrides for our error dataclass enum.