refactor(app): tell TypeScript that aspirateState/dispenseState are part of QuickTransferSummaryState (#20212)
# Overview

This is a type-hinting change to tell TypeScript that the
`aspirateState`/`dispenseState` in `retrieveLiquidClassValues.ts` is
part of `QuickTransferSummaryState`.

Without this, VS Code doesn't find this file when we're looking for
places that touch `.disposalVolumeDispenseSettings`, wasting a lot of
time for programmers trying to find what is setting that field.

Adding this annotation also helps TypeScript catch typos. For example,
if this file had accidentally misspelled the field as
`disposalVolumeDispenseSettingsxx`, TypeScript would now complain about
it.

## Test Plan and Hands on Testing

Just run CI.

## Risk assessment

Very low. No functional change whatsoever.