fix(robot-server): Harmonize error recovery camera field name (#20221)
Closes RQA-4807

To denote image capture on protocol error, the HTTP API uses the
`errorRecoveryCameraEnabled` field except for `GET /camera`, in which
case we use `errorRecoveryEnabled`. The JS code expects the shape of the
camera settings to conform to the pattern elsewhere. To keep the HTTP
API and the app consistent, let's update our one usage of
`errorRecoveryEnabled` to `errorRecoveryCameraEnabled`.

---------

Co-authored-by: Rhyann Clarke <146747548+rclarke0@users.noreply.github.com>