(docs)api: using capture_image in 8.8  (#20200)
# Overview

adds a simple section to Utility Commands to show users how to take an
image (and cover some optional command).

## Test Plan and Hands on Testing

sandbox. 

## Changelog

Added a new section to Utility Commands with some very small updates to
the API reference.

## Review requests

I chose two optional parameters for the code sample that I thought users
would use frequently without getting too complicated, but if we want the
example to be slightly different that's totally fine by me.

Should we link the Flex manual? You can do a lot more with the camera in
8.8 from the app/ODD; but I know we don't usually link out from the API
docs.

## Risk assessment

low.

---------

Co-authored-by: Edward Cormany <edward.cormany@opentrons.com>