fix(protocol-designer): improve logic of eager tip dropping in `generateRobotStateTimeline` (#20246)
"Eager tip dropping" is the term we use to describe handling dropping
tip at the end of a liquid handling step step if 1) the pipette used in
that step is reused, and 2) the first instance in which it is reused
specifies tip handling policy that is not `never`, meaning it will not
require the tips from the pipette's previous step. This allows us to not
arbitrarily keep "dirty" tips on the pipette until its next use if those
tips will not actually be used.