fix(step-generation): use tip rack from previous step if changeTip=never (#20255)
# Overview

This is a temporary Python-only fix for AUTH-2537:

Previously, in step-generation for Transfer steps, we would emit a
liquid class with:
```
pipette.transfer_with_liquid_class(
  liquid_class=protocol.define_liquid_class(
    properties={ ... : {"tip rack the user selected for this step": {
      ...
    }
  )
)
```
But this is wrong if the user wants a tip handling policy of "never" for
this step, because "never" means we need to perform this transfer with
the tip that we picked up from a **previous** step, which could have
come from a different tip rack than the one the user selected for this
step.

This bug would cause the generated protocol to fail analysis.

This PR changes step-generation to use the tip rack from
`robotState.tipState.pipettes[pipette].tiprackURI` when generating
Python for `changeTip=="never"`.

## Test Plan and Hands on Testing

Here is a protocol from PD before this fix, that demonstrates the bug.
In Step 1, it uses `opentrons_96_tiprack_20ul`. In Step 2, it tries to
use `opentrons_96_tiprack_10ul` with tip handling "never". PD
successfully exports the protocol, but it fails in analysis:
[Tip_Management_Never.py](https://github.com/user-attachments/files/23862930/Tip_Management_Never.py)

The protocol will pass analysis when it's re-exported from PD with this
PR.

## Risk assessment

Low-medium. This logic is somewhat tricky.