fix(app): fix pipette card disablement during run (#20264)
Closes RQA-4921

In #20088, we simplified the "robot is busy" banner state in robot settings by using a simplified conditional check, useRobotIsBusy. That works great for the robot settings page, and while we tried to harmonize the "busy" state logic by extending it to the robot details page, that didn't work out as well: we want users to be able to swap/calibrate pipettes on this page while a run is being set up. It's the only way to do so on the OT-2, too, so we definitely can't block it.

This PR effectively reverts the changes made in #20088 to the robot details page.