Docs OT-2 system description revisions (#20125)
# Overview

This adds a System Description section to the revised OT-2 manual. This
section should contain 3 files:

- components
- pipettes
- robot specs

Sandbox:
https://sandbox.docs.opentrons.com/docs-ot2-system-description-revisions/ot-2/system-description/

RTC-859

## Test Plan and Hands on Testing

Have the usual suspects evaluate the content.

## Changelog

When ready, should be 3 files in a System Description section.

## Review requests

Does the text make sense?
Does the organization make sense?
Are the images clear and relevant?

## Risk assessment

Low - moderate due to complexity of changes.

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>