fix(shared-data): Zero out some additional questionable `cornerOffsetFromSlot`s (#20275)
## Overview

These labware had a `cornerOffsetFromSlot.z` of -0.71, which is dubious
because that would put them slightly underground. Since it's the exact
same number as PR #17723 / EXEC-1268, I suspect this was a copy-paste
error and was not intentional.

Closes EXEC-2087.

## Changelog

For these labware:

* `black_96_well_microtiter_plate_lid`
* `corning_96_wellplate_360ul_lid`
* `corning_falcon_384_wellplate_130ul_flat_lid`

I did this:

* Create a new labware version, v2. It's exactly the same as v1 except
`cornerOffsetFromSlot.z` is zeroed out.
* In PAPIv2.28, start loading these new labware versions by default.

## Test Plan and Hands on Testing

I don't have these labware handy to test with, but we can expect the
gripper positions to move up by 0.71 mm now. I expect that to be fine
because they're currently only at ~7mm of ~10mm.