fix(step-generation): implement pickup safety in pickUpTip  (#20284)
In addition to checking external collisions in `pickUpTip` command
creator, we need to check that the tip pickup is valid (i.e., partially
configured pipettes won't pickup too many tips). This is blocked in the
tip selection UI, but the error can arise if saved steps are reordered,
deeming unsuccessful their tip selections at the time of step creation.