feat(app): show QT runs in RecentProtocolRuns (#20295)
With the decision to treat Quick Transfer protocols more like standard
protocols, we want to show them in the desktop app's
`RecentProtocolRuns` component. This also allows us to view and download
any files from the run, which as of now would include photos taken
during any potential error recovery.

There are also a few instances in ODD where we implicitly delete QT
runs. We don't want to do this!