Update labware-creator's ConditionalLabwareRender.
This code was already using a `useLayoutEffect()` to automatically compute an SVG viewbox that enclosed the labware, so it never really mattered how the labware was translated in 2D space. The viewbox would automatically translate to follow it, no matter what. So we can just flip `offsetInSlot` to `passThrough` here without needing to change anything else.