feat(protocol-designer, step-generation): create new manual tip selection modal and timeline errors (#20318)
This PR creates and wires up timeline errors for 1) too many tips picked
up and 2) incomplete pickup in `pickUpTip` command creator according to
designs. Also, it refactors the existing logic for rendering an error inline
notification on saving a bad tip selection, and adds descriptive copy to
the notification depending on the reason for the bad selection.