fix(protocol-designer): document that pipettes.tiprackURI is a labwareId, not a URI (#20326)
# Overview

(This was a small issue I ran into while working on the QT disposal
volume bugs.)

We have this field `RobotState.tipState.pipettes[pipetteId].tiprackURI`.
Despite its name, `.tiprackURI` not a URI, it's a labware entity ID.

This PR adds a docstring explaining that `.tiprackURI` is a labwareId,
and fixes one place in the code that incorrectly assumed it was a URI.

## Test Plan and Hands on Testing

Run CI tests.
Tried it locally with `make -C protocol-designer dev` when using tip
management `never`.

## Risk assessment

Low.