fix(step-generation, shared-data): wire up deck extent checks in safePipetteMovements (#20345)
This PR wires up deck extent checks in `safePipetteMovements.ts` in step
generation. In short, we need to check not only on-deck collisions with
partial tip configuration pipette movements, but also absolute pipette
target positions relative to the limits of the gantry. As part of this
fix, I also add a thermocycler offset to correctly get the well position
for wellplates inside of a thermocycler.