chore(app): shard the frontend unit tests because they take too long (#20341)
# Overview

The app frontend unit tests take > 30 minutes to run nowadays, which
slows me down greatly when I'm waiting for the CI tests to finish. This
change takes advantage of Vitest's sharding functionality (`--shard`) to
split the tests up and run them in parallel on multiple GitHub runners.
https://vitest.dev/guide/improving-performance.html#sharding

The idea is that each shard runs a subset of the frontend unit tests.
Each runner produces a Vitest result file (via `--reporter=blob`) and a
coverage result file. We upload those files to temporary GitHub workflow
artifact storage.

Then another runner waits for all the shards to finish, downloads the
shard results, and merges them into a single Vitest result (with `vitest
--merge-reports`), and merges the coverage results into a single file to
send to Codecov.

This PR also fixes a small bug:

Previously, we were not specifying our Codecov token in the `Upload
coverage report` step, so it was failing with the error `Please upload
with the Codecov repository upload token to resolve issue`, but no one
had noticed the problem before.

## Test Plan and Hands on Testing

Tested manually by carefully watching the CI runs on this PR.

## Risk assessment

Low.