feat: bump our system builders to new versions (#20109)
This is the work-in-progress branch for the monorepo side of bumping
buildroot (https://github.com/Opentrons/buildroot/pull/269 ) and oe-core
( https://github.com/Opentrons/oe-core/pull/237 ) to more recent
versions.

The goal with this work is to only update our direct dependencies as
little as possible to get the system builder updates working. Once the
system builder updates are merged (ideally with little work in this
repo), we can update our direct dependencies to the highest versions
supported by the system builders at their new versions.

For more details on what is changing, see the system builder PRs; the
biggest thing to note here is

**The new python version for both Flex and OT-2 will be 3.12**

## To come out of draft
- [x] oe-core builds
- [x] buildroot builds
- [x] oe-core boots
- [x] buildroot boots
- [x] user installed packages more or less still work

## To merge
- [ ] regression test ot-2 
- [ ] regression test flex

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: sfoster1 <3091648+sfoster1@users.noreply.github.com>
Co-authored-by: Nick Diehl <47604184+ncdiehl11@users.noreply.github.com>
Co-authored-by: Rhyann Clarke <146747548+rclarke0@users.noreply.github.com>
Co-authored-by: Jamey Huffnagle <jamey.huffnagle@opentrons.com>
Co-authored-by: TamarZanzouri <zanzouritamar@gmail.com>
Co-authored-by: Koji Kanao <baxin1919@gmail.com>
Co-authored-by: HovanNgoOpentrons <hovan.ngo@opentrons.com>
Co-authored-by: Ryan Howard <ryan.howard@opentrons.com>
Co-authored-by: Rhyann Clarke <rhyann.clarke@opentrons.com>
Co-authored-by: Josh McVey <josh.mcvey@opentrons.com>