feat(hardware-testing): Create a protocol version of belt calibration (#20281)
# Overview
This protocol does all the same stuff as the previous hardware testing
script except it does not have any dependencies on hardware testing. It
also does not export a CSV report since this is only used in the factory
and this is meant for use in the field during preventative maintenance.