Fix Thermocycler steps sometimes accidentally being created with the block activated.
Some code needed to get updated to account for the new property name. Nothing caught this before because it was obscured by a ts-expect-error.

Also, strangely, this bug only happened for commands other than the very first command in the protocol. The very first command in the protocol is handled specially. That seems wrong, so I've left some todo comments.