chore(system-server): fix test hangs
Tests were hanging (more often in ci, but also locally) because
sometimes the dev server for the tavern backend would fail to start
because a port was in use. Use a random port instead.

Probably better would be letting uvicorn pick which one to use, but it's
a bit of a pain to figure out which one it picked so don't.