refactor(protocol-designer): Delete unused properties in Thermocycler substeps (#20391)
## Overview

In our UI for showing substeps, Thermocycler profiles are peculiar in that we don't actually get data from `ThermocyclerProfileSubstepItem`. Instead, we get it from the step form. This means the properties of `ThermocyclerProfileSubstepItem` are unused by anything.

The properties of `ThermocyclerProfileSubstepItem` would need to update in #20390. Since they're unused, it seems easier to just delete them instead of updating them.

This goes towards EXEC-2009.