refactor(protocol-designer): figure out why SecondStepsMoveLiquidTools is crashing (#20401)
# Overview

`<SecondStepsMoveLiquidTools>` is failing somehow because it can't find
the requested liquid class:

https://opentrons-76.sentry.io/issues/6987511353?project=4509363266387968

The liquid class we want comes from `formData.liquidClass`, so:

- There could be an invalid liquid class name in the form somehow,
- Or it's undefined,
- Or the `NONE_LIQUID_CLASS_NAME` -> `WATER_LIQUID_CLASS_NAME` fallback
is broken?

This PR tries to catch the error with a more informative error message
to help us figure out what's going on.

## Test Plan and Hands on Testing

Smoke-tested locally with a Transfer step.
Run CI tests.

## Risk assessment

Low.