refactor(step-generation): Update the Thermocycler state shape to allow keeping track of ongoing profiles (#20358)
## Overview

This is another step towards EXEC-2009.

On the road to getting step-generation to support concurrent
Thermocycler profiles, this lets it represent the fact that a profile is
currently in progress (i.e., a profile has been started and has not yet
been awaited).

## Changelog

`ThermocyclerModuleState` is expanded:

* The `blockTargetTemp` property is generalized. It's now
`currentBlockActivity`.
  
  Formerly, `blockTargetTemp` could only represent "the block is targeting
a temperature" and "the block is deactivated." It can now additionally
represent "the block is running a Thermocycler profile."
  
  Nothing sets the state to "the block is running a Thermocycler profile"
yet, nor does anything look for it. We're just updating the type to
allow it.

* Not only do we store the fact that a profile is running, but also the
steps of the profile, and the Protocol Engine task ID that it was
started with. In subsequent PRs, this information will help us when we
process `waitForTask` commands and need to update the Thermocycler state
to end the profile.

* A new property, `numProfilesStarted`, counts the number of times a
profile has been started on the Thermocycler. This is going to be used
to generate unique Python variable names, like `task_1 =
thermocycler.start_execute_profile(...)`, `task_2 =
thermocycler_start_execute_profile(...)`, etc.