feat(step-generation): Support concurrent Thermocycler profiles (#20390)
## Overview

This completes the step-generation portion of concurrent Thermocycler
profiles, going towards EXEC-2009.

In #20358, we extended step-generation's data model to allow
representing Thermocycler profiles that are currently in progress. In
this PR, we add some new concurrent commands to step-generation that
actually use that data model.

## Details

**First**, the `thermocyclerProfileStep` compound command has a new
argument, `concurrent: boolean`.
* `concurrent: false` is the preexisting behavior. We'll wait for the
profile to complete before moving on.
* `concurrent: true` emits the new concurrent Thermocycler profile
commands: Protocol Engine's `startRunExtendedProfile`, and the Python
Protocol API's `start_execute_profile()`. We'll start the profile
running in the background and then return immediately.
* Because of backend limitations, "hold steps" (like setting a lid
temperature immediately after the profile ends) are not supported with
`concurrent: true`.

**Second**, there's a new `waitForModuleTask` compound command. You give
it a module ID and your expectation of what the module is currently
doing in the background, like "this Thermocycler is running a profile."
Then it tries to find a handle for that background task. If successful,
it emits a `waitForTask` atomic command to wait for that task to finish.

**So, putting it all together,** the idea is that protocol-designer can
emit sequences of step-generation steps like:

1. `transfer`
2. `thermocyclerProfileStep` with `concurrent: true` and `moduleId:
"abc123"`
3. `transfer`
4. `waitForModuleTask` with `moduleId: "abc123"`
5. `transfer`

And step-generation will convert that into Python like:

```python
pipette.transfer(...)
thermocycler_task_1 = thermocycler.start_execute_profile(...)
pipette.transfer(...)
protocol.wait_for_tasks[thermocycler_task_1]
pipette.transfer(...)
```

This is not quite wired up to protocol-designer yet. That will happen in
another PR.