fix(protocol-designer): add define for lablib (#20442)
Labware library after 8859804da891197d127556b3abb0bbfabdce6748 requires
process.env.NODE_DEBUG to be defined, for some reason; I don't know what
it is, but if it's not then we try and touch it in the browser, which
fails. Define it in PD since PD includes lablib source to fix PD
quickly.