chore: format our ESM files (#20443)
.mts and .mjs were both not in the format file list and present in
eslintignore, for some reason. Let's see what happens if we don't do
that, and automatically format them since they're also code.

This is a format-only PR; there are no code changes.

Note: this is stacked on top of #20442 for better diffing and no
conflicts, but will only get merged into edge once that PR is in