chore: switch mainline python projects to ruff
Ruff ( https://docs.astral.sh/ruff ) is a modern code checking tool that
combines many formatting and style/correctness linting tools and is
really fast. It has a lot of options for things that have previously
been flake8 plugins or separate tools, and it's configured in
pyproject.toml instead of a separate file.

It also has a lot of options for things we've wanted forever but haven't
found a good tool for yet - like import sorting!

Its formatting codestyle philosophy is very similar to black - not a lot
of options, which is good - and so is its formatting output... but its
formatting output is not _identical_ to black's, and so there is a big
old diff here that is 99% automatic and 1% "that type: ignore or noqa is
in the wrong place now".