fix(protocol-designer): stop recording Sentry breadcrumbs for Mixpanel (#20449)
# Overview

Sentry records "breadcrumbs" leading up to an event. The breadcrumbs
include things like the HTTP requests we made before PD crashed. But
because we're using Mixpanel integration, our breadcrumbs are full of
XHR requests to Mixpanel, which are noisy and unhelpful.

The Mixpanel requests are also huge, which could bloat up the Sentry
event and make it too big: see
https://github.com/getsentry/sentry-javascript/issues/874

This PR filters out the breadcrumbs from Mixpanel requests.

## Test Plan and Hands on Testing

Try it?

## Risk assessment

Low.